Every function in its own file, part 8 [#CLICKHOUSE-2]
